sigmoid on float32

+++ fairseq/utils.py
@@ -402,7 +402,7 @@ def softmax(x, dim: int, onnx_trace: bool = False):
     if onnx_trace:
         return F.softmax(x.float(), dim=dim)
     else:
-        return F.softmax(x, dim=dim, dtype=torch.float32) * torch.sigmoid(x)
+        return F.softmax(x, dim=dim, dtype=torch.float32) * torch.sigmoid(x.float())
         # return F.softmax(x, dim=dim, dtype=torch.float32)
